Split out winpty's misc/{buffer-tests,console-handles.md}
This subproject is somewhat interesting independent from winpty: it
explains how the standard handles interact with CreateProcess, which is
something I don't see anywhere else on the Internet.  It's a large body
of code that is functionally independent from winpty, so to keep the
winpty project smaller, it makes sense to split this out.

The code comes from this commit:

  30c6e9f7e0d810a7933ea434b449c77e851603ee
  Generate an import lib to accompany winpty.dll.
  Sun Feb 21 01:35:30 2016 -0600

It looks like the material is at least mostly the same as the 0.2.1
release.  I modified the code (e.g. the Makefile) a bit to get it to
compile.